requirements.txt
numpy==1.19.4
pyyaml==5.3.1
wheel==0.34.2
scipy==1.5.2
pillow==7.2.0
matplotlib==3.3.0
tqdm==4.56.1
tensorboardX==2.0
seaborn==0.10.1
pandas==1.2.0
requests==2.25.0
xvfbwrapper==0.2.9
torchdiffeq==0.2.1
timm==0.5.4
lmdb
Ninja
foolbox
torchsde
statsmodels
transformers
blobfile
mpi4py